Fix casting issues which were making redundanting conversion issues

+++ src/nativeMain/kotlin/maryk/rocksdb/Slice.kt
@@ -20,7 +20,7 @@ actual class Slice actual constructor(
     }
 
     override fun size(): Int {
-        return data.size.toInt()
+        return data.size
     }
 
     override fun empty(): Boolean {
